+++ time_utils.h
@@ -1,6 +1,6 @@
 /*!\file time_utils.h
 ** \author SMFSW
-** \copyright MIT (c) 2017-2018, SMFSW
+** \copyright MIT (c) 2017-2019, SMFSW
 ** \brief Time related utilities
 **/
 /****************************************************************/